move float checks into check module
